Initial commit to update directory skeleton for Paddle frontend
