+++ ivy/functional/frontends/__init__.py
@@ -4,15 +4,17 @@ versions = {
     "numpy": "1.23.2",
     "jax": "0.3.16",
     "scipy": "1.10.1",
+    "paddle": "2.4.2",
 }
 
 import importlib
 
 from . import numpy
 from . import jax
+from . import paddle
+from . import scipy
 from . import tensorflow
 from . import torch
-from . import scipy
 
 
 def fn_name_from_version_specific_fn_name(name, version):
@@ -100,3 +102,4 @@ set_frontend_to_specific_version(tensorflow)
 set_frontend_to_specific_version(jax)
 set_frontend_to_specific_version(numpy)
 set_frontend_to_specific_version(scipy)
+set_frontend_to_specific_version(paddle)
