+++ ivy/functional/frontends/paddle/__init__.py
@@ -1,6 +1,17 @@
 from . import nn
+from . import linalg
+from . import fft
+from . import signal
 
-from .tensor import *
-from .pointwise_ops import *
-from .creation_ops import *
-from .indexing_slicing_joining_mutating_ops import * 
+from .tensor.creation import *
+from .tensor.linalg import *
+from .tensor.logic import *
+from .tensor.manipulation import *
+from .tensor.math import *
+from .tensor.random import *
+from .tensor.search import *
+from .tensor.einsum import *
+
+from .tensor.tensor import Tensor
+
+_frontend_array = Tensor
