+++ ivy/functional/frontends/paddle/linalg.py
@@ -1,0 +1,3 @@
+# All functions are supposed to be added
+# to `paddle.tensor.linalg` namespace and will
+# be imported here for the sake of namespace consistency.
