+++ ivy/functional/frontends/paddle/nn/functional/activation.py
@@ -2,6 +2,7 @@
 import ivy
 from ivy.functional.frontends.paddle.func_wrapper import to_ivy_arrays_and_back
 
+
 @to_ivy_arrays_and_back
-def tanh(input):
-    return ivy.tanh(input)
+def tanh(x, name=None):
+    return ivy.tanh(x)
