+++ ivy/functional/frontends/paddle/tensor/__init__.py
@@ -1,0 +1,9 @@
+from .creation import *
+from .einsum import *
+from .linalg import *
+from .logic import *
+from .manipulation import *
+from .math import *
+from .random import *
+from .search import *
+from .stat import *
