+++ ivy_tests/test_ivy/test_frontends/test_paddle/test_tensor/test_math.py
@@ -1,20 +1,18 @@
 # global
-import ivy
-from hypothesis import assume, strategies as st
 
 # local
 import ivy_tests.test_ivy.helpers as helpers
 from ivy_tests.test_ivy.helpers import handle_frontend_test
 
 
-# tanh
+# sin
 @handle_frontend_test(
-    fn_tree="torch.nn.functional.tanh",
+    fn_tree="paddle.sin",
     dtype_and_x=helpers.dtype_and_values(
         available_dtypes=helpers.get_dtypes("float"),
     ),
 )
-def test_torch_tanh(
+def test_paddle_sin(
     *,
     dtype_and_x,
     on_device,
@@ -29,6 +27,5 @@ def test_torch_tanh(
         test_flags=test_flags,
         fn_tree=fn_tree,
         on_device=on_device,
-        atol=1e-2,
         input=x[0],
     )
